Add signature prefix to eth_sign method
The eth_sign method has been considered insecure for a long time,
because it allows signing arbitrary data blobs, which can leak personal
information, and be used to sign transactions without the usual approval
process.

We have long implemented the preferred alternative as personal_sign,
which is actually not the same as any other client.

For both security and cross-client compatibility, this change brings
eth_sign up to the same behavior as personal_sign, without deprecating
that method, again for retroactive compatibility.

Fixes #2215